+++ src/components/DateTime/DateTimeLabel/DateTimeLabel.css
@@ -1,3 +1,3 @@
-.DateTimeMonthLabel {
+.DateTimeLabel {
   text-transform: capitalize;
 }
