+++ src/components/DateTime/DateTimeToggler/DateTimeToggler.css
@@ -1,4 +1,4 @@
-.DateTimeMonthToggler {
+.DateTimeToggler {
   display: flex;
   justify-content: space-between;
   align-items: center;
