+++ src/components/DateTime/DateTimeTypeYearViewClassic/DateTimeTypeYearViewClassic.css
@@ -1,0 +1,8 @@
+.DateTimeTypeYearViewClassic {
+  --date-time-cell-width: calc(var(--control-height-s) + (var(--space-3xs) * 2));
+  width: calc(var(--date-time-cell-width) * 7);
+
+  &-Label {
+    margin-bottom: var(--space-l);
+  }
+}
