+++ src/utils/date/index.ts
@@ -1,13 +1,7 @@
-import { format } from 'date-fns';
-import ruLocale from 'date-fns/locale/ru';
-
-export const getMonthTitle = (date: Date): string => {
-  return format(date, 'LLLL', { locale: ruLocale });
-};
-
 export * from './isInMinMaxDade';
 export * from './isDateRange';
 export * from './isOnlyOneDateInRange';
 export * from './getCentury';
 export * from './startOfCentury';
 export * from './endOfCentury';
+export * from './isEqualDate';
